runtime: a quick PoC of DedicateOSThread
This is a quick PoC of the DedicateOSThread proposal [1].

`DedicateOSThread()` implies `LockOSThread()`, that is to say,
it locks the invoking G to a M, and also locks the M to a P.

When a normal G calls `DedicateOSThread()`, it will hand off its
normal P, and allocate a new P from the `sched.pidle2` list which
maintains the unused idle Ps reserved for DedicateOSThread.

When a dedicated G needs to pause (e.g. when STW) it will hand
off its dedicated P, and it will be added to `sched.runq` waiting
to be scheduled (this isn't very efficient and can be improved),
and its dedicated M will be stopped by `stoplockedm()`.

When a dedicated P is handed off, it will be added to the
`sched.pidle3` list, and will not be taken by any other Gs/Ms.
And when its dedicated G/M need the P, that P will be taken
from `sched.pidle3` list directly via `pidleget3()`.

When a dedicated G calls `UndedicateOSThread()`, it will release
its dedicated P and return it to the `sched.pidle2` list and
acquire a normal P from the `sched.pidle` list.

There are some limitations/hacks in this PoC, e.g.:

1. GC background mark workers are not created for the Ps reserved
   for DedicateOSThread.
2. `DedicateOSThread()` can not be nested yet.
3. The locked Ps are still counted against `GOMAXPROCS()`.

Currently, the idea behind this PoC is to keep things as simple
as possible, so there is no optimization yet.

[1] https://github.com/google/gvisor/commit/bd4ec85bb10d

Signed-off-by: Tiwei Bie <tiwei.btw@antfin.com>